requirements-dev.txt
Django==3.0.2
djangorestframework==3.11.0
django-picklefield==2.0
pandas==0.25.3
XlsxWriter==1.2.7
matplotlib==3.1.2

# Dev specific dependencies
pylint==2.3.0
git-pylint-commit-hook==2.5.1
pycodestyle==2.5.0
flake8==3.7.6
pep8-naming==0.5.0